+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Fixed Tia.spr() uv-st offset
 - Added entityId to Archetypes
 - Fixed AssetRef opts typing
+- Added constructor options to FlexCanvas
+
+- Removed FlexCanvas.create()
 
 ## v0.0.1
 - Initial release!